requirements.txt
aiogram
curl-cffi
python-dotenv
